+++ test/unit/sonarqube_code_coverage.sh
@@ -8,6 +8,7 @@ coverage erase
 
 echo ""
 echo "Running unit test modules in conjunction with coverage"
+coverage run -a --source=mongo_db_restore test/unit/mongo_db_restore/get_req_options.py
 coverage run -a --source=mongo_db_restore test/unit/mongo_db_restore/help_message.py
 coverage run -a --source=mongo_db_restore test/unit/mongo_db_restore/main.py
 coverage run -a --source=mongo_db_restore test/unit/mongo_db_restore/run_program.py